Cheesy work around for /etc/hosts. Cleaned up and updated postgresql Dockerfile and scripts. Gleaned down the DB in create-db.sh

+++ postgresql/sbin/setup-postgresql.sh
@@ -33,23 +33,9 @@ function usage() {
 	exit 1
 }
 
-function callInclude() {
-# Test for script and run functions
-
-        if [ -f ${WRKDIR}/sbin/include.sh ]; then
-                source ${WRKDIR}/sbin/include.sh
-        else
-                echo "ERROR: The file ${WRKDIR}/sbin/functions not found."
-                echo "This required file provides supporting functions."
-		exit 1
-        fi
-}
 
 function installPostgreSQL() {
 # Install software
-
-	# Pull in rpm for 9.6
-	# RUN rpm -Uvh http://yum.postgresql.org/9.6/redhat/rhel-7-x86_64/pgdg-redhat-repo-latest.noarch.rpm
 	# Pull in rpm for 10
 	yum install https://download.postgresql.org/pub/repos/yum/reporpms/EL-7-x86_64/pgdg-redhat-repo-latest.noarch.rpm
 
@@ -60,40 +46,26 @@ function installPostgreSQL() {
 	/usr/pgsql-10/bin/postgresql10-setup initdb
 }
 
-function quickConfig() {
-
-	echo 'host all all 0.0.0.0/0 md5' >> ${PGDATA}/pg_hba.conf
-	echo "listen_addresses = '*'" >> ${PGDATA}/postgresql.conf
-}
-
 function configPostgreSQL() {
 # Two ways of configuring Postgres. The first is a fast work around.
 # The second is configured to specific databases 
-
-	checkFile ${WRKDIR}/conf/custom_postgresql.conf
-	cp ${WRKDIR}/conf/custom_postgresql.conf ${PGDATA}/postgresql.conf
-	cp ${WRKDIR}/conf/custom_pg_hba.conf ${PGDATA}/pg_hba.conf
-	chown postgres:postgres ${PGDATA}/postgresql.conf
-	chown postgres:postgres ${PGDATA}/pg_hba.conf
+	if [ -f ${WRKDIR}/conf/custom_postgresql.conf ]; then
+		cp ${WRKDIR}/conf/custom_postgresql.conf ${PGDATA}/postgresql.conf
+		cp ${WRKDIR}/conf/custom_pg_hba.conf ${PGDATA}/pg_hba.conf
+		chown postgres:postgres ${PGDATA}/postgresql.conf
+		chown postgres:postgres ${PGDATA}/pg_hba.conf
+	else
+		echo "ERROR: no configuration files in /usr/local/conf" >> ${LOGFILE}
+	fi
 }
 
 function enablePostgreSQL() {
 # Enable Postgresql
-
-	systemctl enable postgresql-10.service
 	systemctl restart postgresql-10.service
-	systemctl status postgresql-10.service &>> /var/log/postgresql-startup.log
+	systemctl status postgresql-10.service &>> ${LOGFILE}
 }
 
 # MAIN
-# Source functions
-#callInclude
-
-# Run checks
-#checkSudo
-
-# Run option
 #installPostgreSQL
-#quickConfig
 configPostgreSQL
 enablePostgreSQL
